Extract playlist header actions into extension
Move PlaylistDetailView header action UI into a new PlaylistDetailView+HeaderActions.swift file and remove the duplicated implementations from PlaylistDetailView.swift. Adjusted access levels (private -> internal) and @Environment/@State properties to allow the extension to use them, and exposed playAll/playableTracks/toggleLibrary for external access. Also fixed qualifier usage (use self instead of Self) in parsers and minor indentation/formatting tweaks in PlayerService+Queue, PlayerBar, and SidebarPinnedItem.
